Reviewed by:	Brian Behlendorf Submitted by:	Chuck Murcko
Fixed -Wall warning.  Woohoo!


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@76957 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/http_protocol.c
@@ -342,8 +342,8 @@ int set_last_modified(request_rec *r, time_t mtime)
      */
 
     if (r->finfo.st_mode != 0)
-        sprintf(weak_etag, "W/\"%lx-%lx-%lx\"", r->finfo.st_ino,
-		r->finfo.st_size, mtime);
+        sprintf(weak_etag, "W/\"%lx-%lx-%lx\"", (unsigned long)r->finfo.st_ino,
+		(unsigned long)r->finfo.st_size, mtime);
     else
         sprintf(weak_etag, "W/\"%lx\"", mtime);
